# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kürsü API",
    "description": "Anonim ders ve akademisyen değerlendirme platformunun API'si. Zarflar, akışlar, maskeleme ve hata gövdeleri gibi bu şemadan okunamayan kurallar için `docs/api-contract.md`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/course-professors": {
      "post": {
        "tags": [
          "course-professors"
        ],
        "summary": "Create Course Professor",
        "operationId": "create_course_professor_course_professors_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseProfessorCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseProfessorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "course-professors"
        ],
        "summary": "List Course Professors",
        "operationId": "list_course_professors_course_professors_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          },
          {
            "name": "term",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Term"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Sayfa başına kayıt",
              "default": 50,
              "title": "Limit"
            },
            "description": "Sayfa başına kayıt"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Atlanacak kayıt sayısı",
              "default": 0,
              "title": "Offset"
            },
            "description": "Atlanacak kayıt sayısı"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CourseProfessorListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/course-professors/{course_professor_id}": {
      "get": {
        "tags": [
          "course-professors"
        ],
        "summary": "Get Course Professor Detail",
        "operationId": "get_course_professor_detail_course_professors__course_professor_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_professor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Professor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseProfessorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/professors": {
      "get": {
        "tags": [
          "professors"
        ],
        "summary": "List Professors",
        "operationId": "list_professors_professors_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hoca adında arama",
              "title": "Search"
            },
            "description": "Hoca adında arama"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Sayfa başına kayıt",
              "default": 50,
              "title": "Limit"
            },
            "description": "Sayfa başına kayıt"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Atlanacak kayıt sayısı",
              "default": 0,
              "title": "Offset"
            },
            "description": "Atlanacak kayıt sayısı"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProfessorListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/professors/{professor_id}": {
      "get": {
        "tags": [
          "professors"
        ],
        "summary": "Get Professor Detail",
        "operationId": "get_professor_detail_professors__professor_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "professor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Professor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Courses",
        "operationId": "list_courses_courses_get",
        "parameters": [
          {
            "name": "department_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bölüm ID (verilmezse search zorunlu)",
              "title": "Department Id"
            },
            "description": "Bölüm ID (verilmezse search zorunlu)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ders adı/kodunda arama",
              "title": "Search"
            },
            "description": "Ders adı/kodunda arama"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Sayfa başına kayıt",
              "default": 50,
              "title": "Limit"
            },
            "description": "Sayfa başına kayıt"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Atlanacak kayıt sayısı",
              "default": 0,
              "title": "Offset"
            },
            "description": "Atlanacak kayıt sayısı"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CourseResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Create Course",
        "operationId": "create_course_courses_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_id}": {
      "patch": {
        "tags": [
          "courses"
        ],
        "summary": "Update Course",
        "operationId": "update_course_courses__course_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Delete Course",
        "operationId": "delete_course_courses__course_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CourseCreate": {
        "properties": {
          "department_id": {
            "type": "integer",
            "title": "Department Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "department_id",
          "name",
          "code"
        ],
        "title": "CourseCreate"
      },
      "CourseProfessorCreate": {
        "properties": {
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "professor_id": {
            "type": "integer",
            "title": "Professor Id"
          },
          "term": {
            "type": "string",
            "minLength": 1,
            "title": "Term"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "professor_id",
          "term"
        ],
        "title": "CourseProfessorCreate"
      },
      "CourseProfessorDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          },
          "course_code": {
            "type": "string",
            "title": "Course Code"
          },
          "professor_name": {
            "type": "string",
            "title": "Professor Name"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "average_teaching_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Teaching Score"
          },
          "average_difficulty_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Difficulty Score"
          },
          "average_fairness_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Fairness Score"
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            },
            "type": "array",
            "title": "Reviews"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_name",
          "course_code",
          "professor_name",
          "term",
          "average_teaching_score",
          "average_difficulty_score",
          "average_fairness_score",
          "review_count",
          "reviews"
        ],
        "title": "CourseProfessorDetail"
      },
      "CourseProfessorListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "professor_name": {
            "type": "string",
            "title": "Professor Name"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "avg_teaching": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Teaching"
          },
          "avg_difficulty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Difficulty"
          },
          "avg_fairness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Fairness"
          }
        },
        "type": "object",
        "required": [
          "id",
          "professor_name",
          "term",
          "avg_teaching",
          "avg_difficulty",
          "avg_fairness"
        ],
        "title": "CourseProfessorListItem"
      },
      "CourseProfessorResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "professor_id": {
            "type": "integer",
            "title": "Professor Id"
          },
          "term": {
            "type": "string",
            "title": "Term"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "professor_id",
          "term"
        ],
        "title": "CourseProfessorResponse"
      },
...
      "CourseResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "department_id": {
            "type": "integer",
            "title": "Department Id"
          },
          "department_name": {
            "type": "string",
            "title": "Department Name"
          },
          "faculty_id": {
            "type": "integer",
            "title": "Faculty Id"
          },
          "faculty_name": {
            "type": "string",
            "title": "Faculty Name"
          },
          "university_id": {
            "type": "integer",
            "title": "University Id"
          },
          "university_name": {
            "type": "string",
            "title": "University Name"
          },
          "university_short_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "University Short Name"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "code",
          "department_id",
          "department_name",
          "faculty_id",
          "faculty_name",
          "university_id",
          "university_name"
        ],
        "title": "CourseResponse"
      },
      "CourseUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "code": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "department_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Id"
          }
        },
        "type": "object",
        "title": "CourseUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_CourseProfessorListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CourseProfessorListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[CourseProfessorListItem]"
      },
      "Page_CourseResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CourseResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[CourseResponse]"
      },
...
      "Page_ProfessorListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProfessorListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ProfessorListItem]"
      },
...
      "ProfessorDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseProfessorSummary"
            },
            "type": "array",
            "title": "Courses"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            },
            "type": "array",
            "title": "Reviews"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "courses",
          "reviews"
        ],
        "title": "ProfessorDetail"
      },
      "ProfessorListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "course_count": {
            "type": "integer",
            "title": "Course Count"
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count"
          },
          "average_teaching_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Teaching Score"
          },
          "average_difficulty_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Difficulty Score"
          },
          "average_fairness_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Fairness Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "course_count",
          "review_count",
          "average_teaching_score",
          "average_difficulty_score",
          "average_fairness_score"
        ],
        "title": "ProfessorListItem"
      },
...
      "ReviewResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_professor_id": {
            "type": "integer",
            "title": "Course Professor Id"
          },
          "teaching_score": {
            "type": "integer",
            "title": "Teaching Score"
          },
          "difficulty_score": {
            "type": "integer",
            "title": "Difficulty Score"
          },
          "fairness_score": {
            "type": "integer",
            "title": "Fairness Score"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_professor_id",
          "teaching_score",
          "difficulty_score",
          "fairness_score",
          "comment",
          "status",
          "created_at"
        ],
        "title": "ReviewResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}